+++ src/Diffusion/NoisePredictors/AsymmDiTPredictor.cs
@@ -73,7 +73,7 @@ public class AsymmDiTPredictor<T> : NoisePredictorBase<T>
     /// <inheritdoc />
     public override int ContextDimension => _contextDim;
     /// <inheritdoc />
-    public override int ParameterCount { get; }
+    public override long ParameterCount { get; }
 
     /// <summary>
     /// Initializes a new AsymmDiT predictor.
@@ -119,9 +119,9 @@ public class AsymmDiTPredictor<T> : NoisePredictorBase<T>
 
     private int CalculateParameterCount()
     {
-        int count = _patchEmbed.ParameterCount;
-        foreach (var block in _blocks) count += block.ParameterCount;
-        count += _finalLayer.ParameterCount;
+        int count = (int)((int)_patchEmbed.ParameterCount);
+        foreach (var block in _blocks) count += (int)((int)block.ParameterCount);
+        count += (int)((int)_finalLayer.ParameterCount);
         return count;
     }
 
@@ -171,7 +171,7 @@ public class AsymmDiTPredictor<T> : NoisePredictorBase<T>
 
     private static int SetParams(DenseLayer<T> layer, Vector<T> parameters, int offset)
     {
-        int count = layer.ParameterCount;
+        int count = (int)((int)layer.ParameterCount);
         var p = new T[count];
         for (int i = 0; i < count; i++) p[i] = parameters[offset + i];
         layer.SetParameters(new Vector<T>(p));
