+++ src/Diffusion/SuperResolution/SUPIRModel.cs
@@ -193,7 +193,7 @@ public class SUPIRModel<T> : LatentDiffusionModelBase<T>
     public override int LatentChannels => LATENT_CHANNELS;
 
     /// <inheritdoc />
-    public override int ParameterCount => _unet.ParameterCount + _vae.ParameterCount;
+    public override long ParameterCount => _unet.ParameterCount + _vae.ParameterCount;
 
     #endregion
 
@@ -472,8 +472,8 @@ public class SUPIRModel<T> : LatentDiffusionModelBase<T>
             Name = "SUPIR",
             Version = "1.0",
             Description = "SUPIR SDXL-based photo-realistic image restoration with GPT-guided conditioning",
-            FeatureCount = ParameterCount,
-            Complexity = ParameterCount
+            FeatureCount = (int)ParameterCount,
+            Complexity = (int)ParameterCount
         };
 
         metadata.SetProperty("architecture", "sdxl-restoration");
